[FIX/#71] SaveBdayView 리펙토링 완료
